+++ src/application/productos/producto-delete.service.ts
@@ -4,9 +4,19 @@ export class ProductoDeleteService {
   constructor(private unitOfWork: IUnitOfWorkInterface) {}
 
   async execute(id: string) {
-    const productoRepository = this.unitOfWork.productoRepository;
+    const { productoRepository, itemRepository } = this.unitOfWork;
     const producto = await productoRepository.findOne({ where: { id } });
     if (!producto) throw new Error('El regístro no existe!');
-    return productoRepository.delete(producto.id);
+    const count = await itemRepository
+      .createQueryBuilder()
+      .andWhere(`productoId = '${producto.id}'`)
+      .getCount();
+    if (count) {
+      await itemRepository.update(producto.id, { state: false });
+      return { message: 'Registro desactivado' };
+    } else {
+      await productoRepository.delete(producto.id);
+      return { message: 'Registro eliminado' };
+    }
   }
 }
