+++ src/application/situacion/situacion-delete.service.ts
@@ -6,10 +6,20 @@ export class SituacionDeleteService implements IBaseServiceInterface {
   constructor(private unitOfWork: IUnitOfWorkInterface) {}
 
   async execute(params: SituacionFindRequest) {
-    const { situacionRepository } = this.unitOfWork;
+    const { situacionRepository, notaRepository } = this.unitOfWork;
     await this.unitOfWork.start();
     const situacion = await situacionRepository.findOne({ where: params });
     if (!situacion) throw new Error('No se encontró la situación');
-    return situacionRepository.delete(situacion);
+    const count = await notaRepository
+      .createQueryBuilder()
+      .andWhere(`situacionId = '${situacion.id}'`)
+      .getCount();
+    if (count) {
+      await situacionRepository.update(situacion.id, { state: false });
+      return { message: 'Registro desactivado' };
+    } else {
+      await situacionRepository.delete(situacion);
+      return { message: 'Registro elminado' };
+    }
   }
 }
